requirements.txt
Flask>=3.0.0
Flask-Migrate>=4.1.0
Flask-SQLAlchemy>=3.1.0
gpt4all>=2.0.0
PyMySQL>=1.1.0
Werkzeug>=3.0.0
